feat: create new interrupt file
