+++ switch_service.c
@@ -2,6 +2,7 @@
 #include "utils.h"
 #include "switch_service.h"
 #include "led_service.h"
+#include "interrupt_service.h"
 #include "MKL25Z4.h" 
 
 void switch_button_init(void) {
@@ -26,14 +27,3 @@ void switch_button_init(void) {
 }
 
 
-void PORTD_IRQHandler() {
-	// Clear Pending IRQ
-	NVIC_ClearPendingIRQ(PORTD_IRQn);
-	
-	if (PORTD->ISFR & MASK(SWITCH_BUTTON_PIN)) {
-			led_color = led_control(led_color);
-	}
-	
-	// Clear INT Flag
-	PORTD->ISFR |= MASK(SWITCH_BUTTON_PIN);
-}
